Add note's title in the UI
This is a very heavy commit with lots of changes. Let's start with the
small ones. The TextStyle styles now have the lineHeight attribute as it
will have no effect if we put it in TextAppearance.

Now, the user can add a Title to the note as well. It is completely
optional and will only be shown in the list if it is present.